Move most tests to common_test.rs
